refactor(website): Moves mapJson() to playground component Moves this function out of transformation.ts so that we can use the transformation utilities without loading all schemas with Vite
