+++ website/src/lib/transformation.ts
@@ -1,5 +1,4 @@
-import type { JsonValue, FormData, TransformOutput } from "./types";
-import { schemas } from "./schemas";
+import type { JsonValue } from "./types";
 
 /**
  * This module provides a utility function for transforming data using a mapping.
@@ -207,43 +206,4 @@ export function transformWithMapping(
 
   // Recursively walk the mapping until all nested transformations are applied
   return transformNode(mapping, depth) as Record<string, JsonValue>;
-}
-
-/**
- * Maps JSON data from one form schema to another using the common data format as an intermediary.
- *
- * @param data - The source form data
- * @param sourceId - The ID of the source form schema
- * @param targetId - The ID of the target form schema
- * @returns A TransformOutput object containing the transformation results
- */
-export function mapJson(
-  data: FormData,
-  sourceId: string,
-  targetId: string,
-): TransformOutput {
-  const sourceSchema = schemas[sourceId];
-  const targetSchema = schemas[targetId];
-
-  if (!sourceSchema || !targetSchema) {
-    throw new Error("Source or target schema not found");
-  }
-
-  const commonData = transformWithMapping(
-    data,
-    sourceSchema.mappingToCommon as FormData,
-  );
-
-  const targetData = transformWithMapping(
-    commonData,
-    targetSchema.mappingFromCommon as FormData,
-  );
-
-  return {
-    timestamp: Date.now(),
-    source: sourceId,
-    target: targetId,
-    commonData,
-    targetData,
-  };
 }
